[BUGFIX] fix broken option "ifTranslated" for normal tables
If a parent record has "ifTranslated" select a translation is supposed
to show up if (and only if) a translation in the current language is set.
It should not fall back to any other language or the default language.

This patch reintroduces this feature. Also some unit-tests where added
that do not only check the services of languagevisibility but also the
hooks the extension uses.

Change-Id: Icfc29e8ddfa44f5b50b683fecf2aa37b63c27b98
Reviewed-on: https://review.typo3.org/23103
Reviewed-by: Christian Zenker
Tested-by: Christian Zenker
Reviewed-by: Tolleiv Nietsch
Tested-by: Tolleiv Nietsch