[BUGFIX] ignore deleted page_language_overlay records in BE
Deleted page_language_overlay records in the backend where not ignored
when calculating the visibility of a page translation. This could lead to
having wrong visibility data shown in the (non-deleted) overlay record
that will be stored when modifying and saving the record.

The bugfix modifies the query in the BE to not consider deleted records.

Change-Id: Ib3b3d38d8f9321a257c7ff98865b1b72dff56b45
Reviewed-on: https://review.typo3.org/23193
Reviewed-by: Christian Zenker
Tested-by: Christian Zenker
Reviewed-by: Tolleiv Nietsch
Tested-by: Tolleiv Nietsch