Fix of Warning: Illegal offset type in isset (#48)
* Fix of Warning: Illegal offset type in isset

Even though it would be better to use PHP7.1 type hints and rely on $id that is always integer. I'd like to propose this change to fix php warning error. This change also fixes bug of intval($id) (line: 165), because intval of array is null and improves performance.

* Added --prefer-source to composer require

Test classes are not part of TYPO3 version greater than 7.6.11 by default. It requires --prefer-source